fix(ui): snap a slider to the value its label shows
## What this changes

A slider quantised in float, and a step of 0.05 is not a float: nine of
them summed lands on the float above the one that spells 0.45. Three of
the twenty-one stops on the volume slider were therefore one float out,
and the settings file recorded 0.45000002 for a slider reading 45 % --
both true, about different numbers.

The arrow keys did not quantise at all. They added the step to whatever
the value already was, so holding one down walked the value off the grid
a rounding error at a time and wrote every one of them to the file.

Both routes now go through ui::snapToStep, which works in double and
narrows once. All twenty-one volume stops, all nine threshold stops and
all seven opacity stops are now the float nearest the decimal the label
prints.

## Notes for review

The window opacity floor became a double for the same reason. Its slider
counts from that value, and a grid counted from the float nearest 0.7
puts its fourth stop on 0.84999996; counted from the decimal, the stop
is 0.85. It is narrowed at the one place a float is what is wanted.

snapToStep is in PowerPeekCore rather than inside the widget so the grid
can be tested at all -- Widgets wants a device and a window, and this is
arithmetic. One of its cases deliberately reaches across to json::Value
and asserts the text a stop is written as, because that is the promise
the change exists to keep and neither half demonstrates it alone.

This is the tail of the float-in-the-settings-file work: the writer was
made honest first, and this makes the values worth being honest about.

+++ src/core/Settings.cpp
@@ -305,8 +305,9 @@ Settings Settings::load(std::filesystem::path const& file) {
         parseEnum(kToastPositionNames, root["toastPosition"], settings.toastPosition);
 
     settings.backdrop = parseEnum(kBackdropNames, root["backdrop"], settings.backdrop);
-    settings.windowOpacity = std::clamp(root["windowOpacity"].asFloat(settings.windowOpacity),
-                                        kMinimumWindowOpacity, 1.0f);
+    settings.windowOpacity =
+        std::clamp(root["windowOpacity"].asFloat(settings.windowOpacity),
+                   static_cast<float>(kMinimumWindowOpacity), 1.0f);
 
     settings.masterVolume =
         std::clamp(root["masterVolume"].asFloat(settings.masterVolume), 0.0f, 1.0f);
